Implement IEditor as singledispatch function
The editor protocol is implemented as ABC.
